Move Builder Class to it's own file and Save root_dir_abs
